+++ front/app/__libs__/components/ContextMenu/ContextMenu.tsx
@@ -1,38 +1,15 @@
-import { useWebSocket } from "@akasha-utils/react/websocket-hook";
 import { ContextMenuBase } from "./ContextMenuBase";
-import { ByteBuffer } from "@akasha-lib";
-import { useAtomValue, useSetAtom } from "jotai";
+import { useAtomValue } from "jotai";
 import { TargetedAccountUUIDAtom } from "@atoms/AccountAtom";
-import { ChatServerOpcode } from "@common/chat-opcodes";
 import { useProtectedProfile } from "@hooks/useProfile";
 import { useEffect, useRef, useState } from "react";
-import {
-    ActiveStatus,
-    ActiveStatusNumber,
-    RoleNumber,
-    getActiveStatusNumber,
-} from "@common/generated/types";
-import { logoutAction } from "@/app/(main)/@home/(nav)/logoutAction";
-import {
-    useChatMember,
-    useSetChatRightSideBarCurrrentPageAtom,
-} from "@hooks/useChatRoom";
+import { RoleNumber } from "@common/generated/types";
+import { useChatMember } from "@hooks/useChatRoom";
 import {
     useCurrentAccountUUID,
     useCurrentChatRoomUUID,
 } from "@hooks/useCurrent";
-import {
-    makeActiveStatusManualRequest,
-    makeAddEnemyRequest,
-    makeAddFriendRequest,
-    makeChangeMemberRoleRequest,
-    makeDeleteFriendRequest,
-    makeHandoverRoomOwnerRequest,
-    makeModifyFriendRequest,
-} from "@akasha-utils/chat-payload-builder-client";
-import { CurrentChatRoomUUIDAtom } from "@atoms/ChatAtom";
-import { GlobalStore } from "@atoms/GlobalStore";
-import { makeDirectChatKey } from "@akasha-utils/idb/chat-store";
+import { useContextMenuActions } from "./useContextMenuActions";
 
 export type Relationship = "myself" | "friend" | "stranger";
 
@@ -251,12 +228,15 @@ export function ContextMenu({ type }: { type: Scope }) {
     const targetAccountUUID = useAtomValue(TargetedAccountUUIDAtom);
     const profile = useProtectedProfile(targetAccountUUID);
     const currentChatRoomUUID = useCurrentChatRoomUUID();
-    const setCurrentChatRoomUUID = useSetAtom(CurrentChatRoomUUIDAtom, {
-        store: GlobalStore,
-    });
     const currentAccountUUID = useCurrentAccountUUID();
     const currentUser = useChatMember(currentChatRoomUUID, currentAccountUUID);
     const roleLevel = Number(currentUser?.role ?? 0);
+    const actions = useContextMenuActions(
+        targetAccountUUID,
+        currentAccountUUID,
+        currentChatRoomUUID,
+        profile,
+    );
 
     const relationship =
         targetAccountUUID === currentAccountUUID
@@ -265,107 +245,14 @@ export function ContextMenu({ type }: { type: Scope }) {
             ? "friend"
             : "stranger";
 
-    // TODO: profile undefined 면 뭘 어떻게 해야??
-    const { sendPayload } = useWebSocket("chat", []);
-
-    // TODO: 이거 fallback 처리를 여기서 하는게 맞는가,
-    // 아니면 각각 action 함수 내부마다 따로 하나씩 해주는게 맞는가
-    const nickName = profile?.nickName ?? "fallback";
-    const nickTag = profile?.nickTag ?? 0;
-
     //TODO: fetch score
     const score = 1321;
-    const setCurrentPage = useSetChatRightSideBarCurrrentPageAtom();
     const rating: ProfileMenu = {
         name: `rating: ${score}`,
         relation: ["myself", "friend", "stranger"],
         scope: "ChatRoom",
         isImportant: false,
         className: "hover:bg-transparent active:bg-transparent",
-    };
-
-    const actions = {
-        ["copytag"]: () => {
-            // TODO: 복사되었다고 텍스트 바꾸기? setTimeout?
-            navigator.clipboard
-                .writeText(`${nickName}#${nickTag}`)
-                .then(() => {})
-                .catch(() => {});
-        },
-        ["changeactivestatus"]: () => {
-            // TODO: 입력으로 받기
-            const newActiveStatus = ActiveStatus.INVISIBLE;
-            const buf = makeActiveStatusManualRequest(
-                getActiveStatusNumber(newActiveStatus),
-            );
-            sendPayload(buf);
-        },
-        ["addfriend"]: () => {
-            // TODO: true: add friend by nick + tag, false: by id
-            const buf = makeAddFriendRequest(targetAccountUUID, "", 0b11111111);
-            sendPayload(buf);
-        },
-        ["modifyfriend"]: () => {},
-        ["logout"]: () => {
-            // TODO: 현재 이 함수 (nav)폴더에 있는데 어디로 옮기지? util? 아니면 여기 이 폴더?
-            logoutAction();
-        },
-        ["gotoprofile"]: () => {
-            window.open(`/profile/${nickName}/${nickTag}`);
-        },
-        ["addenemy"]: () => {
-            if (confirm(`진짜로 정말로 [${nickName}]님을 차단하실건가요...?`)) {
-                const reason = prompt("왜요...?") ?? "그냥";
-                const buf = makeAddEnemyRequest(targetAccountUUID, reason);
-                sendPayload(buf);
-            }
-        },
-        ["reportuser"]: () => {
-            setCurrentPage("report");
-        },
-        ["directmessage"]: () => {
-            setCurrentChatRoomUUID(
-                makeDirectChatKey(currentAccountUUID, targetAccountUUID),
-            );
-        },
-        ["accessban"]: () => {
-            setCurrentPage("newAccessBan");
-        },
-        ["sendban"]: () => {
-            setCurrentPage("newSendBan");
-        },
-        ["transfer"]: () => {
-            if (confirm("정말로 방을 양도하시겠습니까?")) {
-                sendPayload(
-                    makeHandoverRoomOwnerRequest(
-                        currentChatRoomUUID,
-                        targetAccountUUID,
-                    ),
-                );
-            }
-        },
-        ["grant"]: () => {
-            if (confirm("매니저 지정하시겠습니까?")) {
-                const targetRole = RoleNumber.MANAGER;
-                sendPayload(
-                    makeChangeMemberRoleRequest(
-                        currentChatRoomUUID,
-                        targetAccountUUID,
-                        targetRole,
-                    ),
-                );
-            }
-        },
-        ["deletefriend"]: () => {
-            if (
-                confirm(
-                    `진짜로 정말로 [${nickName}]님을 친구 목록에서 삭제하실건가요...?`,
-                )
-            ) {
-                const buf = makeDeleteFriendRequest(targetAccountUUID);
-                sendPayload(buf);
-            }
-        },
     };
 
     return (
